requirements.txt
Django==3.0.3
psycopg2-binary==2.8.5
pytest==5.3.1
djangorestframework==3.11.0
pytest==5.3.1
pytest-cov==2.8.1
pytest-django==3.9.0
coverage==4.5.4